requirements.txt
alembic==1.7.5
amqp==5.0.9
anyio==3.4.0
appdirs==1.4.4
asgiref==3.4.1
attrs==21.4.0
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
billiard==3.6.4.0
black==21.6b0
boto3==1.21.3
botocore==1.24.3
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
country-converter==0.7.4
coverage==6.3.1
cryptography==36.0.1
distlib==0.3.4
dnspython==2.2.0
docx2txt==0.8
ecdsa==0.17.0
email-validator==1.1.3
factory-boy==3.2.0
Faker==13.0.0
fastapi==0.70.1
fastapi-pagination==0.9.1
FastAPI-SQLAlchemy==0.2.1
filelock==3.4.0
flower==1.0.0
greenlet==1.1.2
h11==0.12.0
humanize==4.0.0
idna==2.9
inflection==0.5.1
iniconfig==1.1.1
jmespath==0.10.0
jwt==1.3.1
kombu==5.2.2
Mako==1.1.6
MarkupSafe==2.0.1
more-itertools==8.12.0
mypy-extensions==0.4.3
numpy==1.22.2
packaging==21.3
pandas==1.4.1
passlib==1.7.4
pathspec==0.9.0
pipenv==2021.11.23
platformdirs==2.4.0
pluggy==0.13.1
prometheus-client==0.13.1
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pycountry==22.1.10
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
redis==3.5.3
regex==2021.11.10
requests==2.25.1
rsa==4.8
s3transfer==0.5.1
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
toml==0.10.2
tornado==6.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
vine==5.0.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
watchgod==0.7
wcwidth==0.2.5
Werkzeug==2.0.2
